Extract ip tools. Fix input data verification errors
